requirements.txt
aiogram>=3.0.0
python-dotenv~=1.1.0
asyncpg~=0.30.0
gspread~=6.2.0
google-api-python-client~=2.166.0
google-auth
tenacity~=9.1.2

protobuf~=6.30.2